Rename interface to `Stream`, implementation to `Streaming`
